[Engine/Bindings/Release/CI] Close the independent M1 exit audit (#29-#69)
One hardening pass over every finding of the Codex M1 exit audit.

Engine: 64-bit scaled points end long-input overflow (UBSan-clean now,
with a public-API regression test and allocation instrumentation);
portable negative fixed-point scaling; lexical style scope stamped at
delivery so an inner math alphabet always wins (new corpus case
math-styles-inner-wins, 80 -> 81); txc_field becomes a tagged union
(items 432 -> 216 bytes) with one full-initialization reset; parser
scratch lives in its own arena released before compile returns; arena
chunks grow geometrically; the CLI reports short writes and flush
failures; the command dispatch table gains a full name/payload
consistency check; threads cover concurrent reads of one shared tree.

Bindings: TXC1 moves to packages/tex-core/bridge with an exact
single-allocation writer; Kotlin/ES/Swift decoders fail closed on
unknown discriminants, non-scalar codepoints, leaf children, count
mismatches, and trailing bytes (with handcrafted malformed-payload
tests); ES decodes from live WASM memory and deep-freezes every public
path; Kotlin HBox snapshots children across construction and copy,
Document.compile gains @JvmStatic, production coroutines dependency
removed, jdk-release=17; Swift borrows contiguous UTF-8.

Release/CI: byte-reproducible C tar and Swift zip (double-staging
proof in the dry run); relocatable pkg-config plus a static contract
carrying TEX_CORE_STATIC_DEFINE, verified from the extracted archive;
LICENSE ships in the C tar and every Maven artifact; macOS deployment
target 15 pinned and vtool-audited; pre-1.0 SONAME 0.minor and
SameMinorVersion with a documented ABI policy; CMake hardening for
embedding, no-library configs, and the 3.21 preset minimum; sanitizer
lanes run the conformance corpus; a seeded, dictionary-guided,
postcondition-checking fuzz campaign joins required CI; Kotlin
consumers verify the exact staged publication artifacts; Node 20 and
JRE 17 minimum-runtime lanes; actions pinned to commit SHAs with
hash-locked installer inputs and a policy audit; CodeQL traces the FFI
bridges and Android sources; C benchmarks join PR metrics with
separate compile/dump boundaries; the repository health check stops
compiling products; bootstrap defaults preserve active protections;
init-environment --check is offline and read-only; Central Portal
calls and publish jobs get hard timeouts; resume-publish binds the
deployment id and serializes on the release tag; docs reconciled with
the completed M1/schema-v5 state.

+++ packages/tex-core/core/scaled.h
@@ -8,7 +8,12 @@
 #include <stddef.h>
 #include <stdint.h>
 
-typedef int32_t txc_scaled;
+/* 64-bit so unbounded accumulation over a whole document cannot overflow:
+ * every leaf measure is under 2^24 sp (a few ems) and every accumulated
+ * term costs at least one parsed item (hundreds of arena bytes), so a sum
+ * would need petabytes of input to approach 2^63. 32 bits overflowed on a
+ * 64 KiB single-line document. */
+typedef int64_t txc_scaled;
 
 /* The text-size em: 10 pt. The script and scriptscript sizes (7 pt and
  * 5 pt) are the layout module's TXC_MATHSIZE ems. */
